+++ src/sql/blocksLatest.sql
@@ -2,12 +2,12 @@ SELECT
 block_hash AS "blockId",
 prev_block_hash AS "prevBlockId",
 block_hashmerkleroot AS "merkleRootHash",
-block_nonce::bigint AS "nonce",
-block_version::bigint AS "version",
-block_height::bigint AS "blockHeight",
-block_bits::bigint AS "blockSize",
-block_timestamp::bigint AS "timestamp",
-block_tx_count::bigint AS "txCount"
+block_nonce AS "nonce",
+block_version AS "version",
+block_height AS "blockHeight",
+block_bits AS "blockSize",
+block_timestamp AS "timestamp",
+block_tx_count AS "txCount"
 FROM block_view
 WHERE block_in_longest = true
 ORDER BY block_height DESC
